% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebma.R
\name{ebma}
\alias{ebma}
\title{Create an EBMA ensemble}
\usage{
ebma(y, x, model_type, y_test = NULL, x_test = NULL,
  model_names = NULL, ...)
}
\arguments{
\item{y}{Outcome values for training data (calibration).}

\item{x}{Input forecasts for training data (calibration).}

\item{model_type}{Model to use, either "normal" or "logit".}

\item{y_test}{Test period outcomes.}

\item{x_test}{Test period input forecasts.}

\item{model_names}{Vector of model names; if NULL, column names of
\code{x} are used.}

\item{\dots}{Additional arguments for \code{\link[EBMAforecast:calibrateEnsemble]{EBMAforecast::calibrateEnsemble()}}}
}
\description{
Wrapper around \code{\link[EBMAforecast:makeForecastData]{EBMAforecast::makeForecastData()}} and \code{\link[EBMAforecast:calibrateEnsemble]{EBMAforecast::calibrateEnsemble()}}.
that mimics R's conventional model syntax.
}
