% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebma.R
\name{predict.ebma}
\alias{predict.ebma}
\title{Predict method for EBMA}
\usage{
\method{predict}{ebma}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class "ebma"}

\item{newdata}{a data frame containing inputs to use for the prediction;
otherwise predictions are for the validation period (in-sample)}

\item{\dots}{not used}
}
\description{
Predict method for EBMA
}
