% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebma.R
\name{predict_FDatFitLogit}
\alias{predict_FDatFitLogit}
\title{Replacement for broken predictor}
\usage{
predict_FDatFitLogit(object, newdata, ...)
}
\arguments{
\item{object}{class "ebma"}

\item{newdata}{optional new data}

\item{\ldots}{not used}
}
\description{
Replacement for broken predictor
}
\keyword{internal}
