% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebma.R
\name{print.ebma}
\alias{print.ebma}
\title{Print method}
\usage{
\method{print}{ebma}(object, ...)
}
\arguments{
\item{object}{Fitted "ebma" object.}

\item{\dots}{Additional arguments.}
}
\description{
Wrapper for EBMAforecast's print method
}
