% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebma.R
\name{summary.ebma}
\alias{summary.ebma}
\title{Summary method}
\usage{
\method{summary}{ebma}(object, ...)
}
\arguments{
\item{object}{Fitted "ebma" object.}

\item{\dots}{Additional arguments for \code{\link[EBMAforecast]{`SummaryForecastData-class`}}}
}
\description{
Wrapper for \code{\link[EBMAforecast]{`SummaryForecastData-class`}}
}
